# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "タスみる API",
    "description": "仕様書からタスクを自動生成し、担当候補を提案するAPI",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/teams": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Create Team",
        "description": "チームを作成し、作成者を管理者メンバーとして登録した上でセッションを発行する",
        "operationId": "create_team_api_teams_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/teams/{team_id}/invitations": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Create Invitation",
        "description": "招待トークンを発行する（対象チームの管理者のみ実行できる）",
        "operationId": "create_invitation_api_teams__team_id__invitations_post",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Team Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/invitations/{token}/join": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Join Team",
        "description": "招待トークンでチームに参加し、新しいメンバーとデバイスセッションを作成する",
        "operationId": "join_team_api_invitations__token__join_post",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinTeamRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/teams/{team_id}/members": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "List Members",
        "description": "チームメンバー一覧を取得する（自チーム以外は404）",
        "operationId": "list_members_api_teams__team_id__members_get",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Team Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamMemberResponse"
                  },
                  "title": "Response List Members Api Teams  Team Id  Members Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/members/{member_id}": {
      "delete": {
        "tags": [
          "auth"
        ],
        "summary": "Delete Member",
        "description": "チームメンバーを削除する（管理者のみ・自チームのメンバーのみ）",
        "operationId": "delete_member_api_members__member_id__delete",
        "parameters": [
          {
            "name": "member_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Member Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InvitationCreateRequest": {
        "properties": {
          "expires_in_hours": {
            "type": "integer",
            "maximum": 720.0,
            "minimum": 1.0,
            "title": "Expires In Hours",
            "description": "招待リンクの有効時間",
            "default": 24
          }
        },
        "type": "object",
        "title": "InvitationCreateRequest"
      },
      "InvitationCreateResponse": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token",
            "description": "招待トークン（この応答でのみ生の値が返される）"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "token",
          "expires_at"
        ],
        "title": "InvitationCreateResponse"
      },
...
      "JoinTeamRequest": {
        "properties": {
          "display_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Display Name",
            "description": "参加者の表示名"
          }
        },
        "type": "object",
        "required": [
          "display_name"
        ],
        "title": "JoinTeamRequest"
      },
      "MeResponse": {
        "properties": {
          "team": {
            "$ref": "#/components/schemas/TeamResponse"
          },
          "member": {
            "$ref": "#/components/schemas/TeamMemberResponse"
          }
        },
        "type": "object",
        "required": [
          "team",
          "member"
        ],
        "title": "MeResponse"
      },
...
      "TeamCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "description": "チーム名"
          },
          "admin_display_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Admin Display Name",
            "description": "作成者（管理者）の表示名"
          }
        },
        "type": "object",
        "required": [
          "name",
          "admin_display_name"
        ],
        "title": "TeamCreateRequest"
      },
      "TeamMemberResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "is_admin": {
            "type": "boolean",
            "title": "Is Admin"
          }
        },
        "type": "object",
        "required": [
          "id",
          "display_name",
          "is_admin"
        ],
        "title": "TeamMemberResponse"
      },
      "TeamResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "TeamResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}